rewrote and completed the lexer, removed Token's From<&str> impl
Rewrote Lexer to make it easier to support symbols longer than one
character (e.g. '==').
Added support for remaining tokens to the lexer.
Removed Token's From<&str> impl because this logic depends upon the
context (previous and next characters, e.g. to check if the next
character is a space), so it belongs more to Lexer than to the token
module.